ai_text_detector/requirements.txt
Django==4.2.4
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0
joblib==1.4.2
scikit-learn==1.5.1
djangorestframework==3.15.2
numpy==1.26.4
psycopg2-binary==2.9.9